{"schema":"decodex/commit/1","summary":"Modularize execution program core","authority":"manual"}
